fix: show loading indicator also during file uploading
ref #2318
